feat: divide datasource from VC
- 뷰컨으로 부터 데이터소스를 분리
